backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
pymongo==4.3.3
motor==3.1.2
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
beautifulsoup4==4.12.2
requests==2.30.0
aiohttp==3.8.4
nltk==3.8.1
scikit-learn==1.2.2
numpy==1.24.3
pandas==2.0.1
transformers==4.28.1
sentence-transformers==2.2.2
python-dateutil==2.8.2
